iceberg-tests/requirements.txt
PyYAML==6.0.2
pydantic==2.8.2
requests==2.32.3
azure-identity==1.17.1
azure-storage-blob==12.22.0
snowflake-connector-python==3.16.0
snowflake-snowpark-python==1.38.0
pyspark==3.5.1
faker==26.0.0
python-dotenv==1.0.1
cryptography==41.0.7
sqlparse==0.5.1
Jinja2==3.1.4
databricks-sql-connector==2.9.3
